rename bench script
