+++ benchmarks/sql/tests/unit/test_main_evaluate.py
@@ -1,8 +1,9 @@
 from unittest.mock import call, patch
 
 import pytest
-from evaluate import evaluate
 from omegaconf import DictConfig
+
+from benchmarks.sql.bench import evaluate
 
 
 @patch("dbally_benchmark.evaluate.e2e_evaluate")
